% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addDataToDB}
\alias{addDataToDB}
\title{Parse data from a combined format and insert it into the database}
\usage{
addDataToDB(combined_data, dbInfo)
}
\arguments{
\item{combined_data}{Dataframe of the original data}

\item{dbInfo}{connection or path to an SQLite database}
}
\value{
TRUE if success
}
\description{
Parse data from a combined format and insert it into the database
}
