% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbAddEvaluations}
\alias{dbAddEvaluations}
\title{Parse data from a combined format and insert it into the database}
\usage{
dbAddEvaluations(combined_data, dbPath, redactedOnly = F)
}
\arguments{
\item{combined_data}{Dataframe of the original data}

\item{dbPath}{Path to a (new) CFME database}

\item{redactedOnly}{(Default = FALSE) If TRUE, only redacted evlaluations are
put into the database the version with identifiers is omitted}
}
\value{
TRUE if success
}
\description{
Parse data from a combined format and insert it into the database
}
