% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbAddLLMreview}
\alias{dbAddLLMreview}
\title{Add an LLM response from llm_review() to the database}
\usage{
dbAddLLMreview(dbInfo, llm_review)
}
\arguments{
\item{dbInfo}{Database info}

\item{llm_review}{Output of the llm_review() function}
}
\value{
A data frame with the following columns:
\itemize{
\item evaluation_id,
\item review_assignment_id,
\item review_score_id: the ID for the each detected competency review scores
}
}
\description{
Add an LLM response from llm_review() to the database
}
