% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbAIreview}
\alias{dbAIreview}
\title{Add an AI response from llm_review() to the database}
\usage{
dbAIreview(dbInfo, llmReview, commit = T)
}
\arguments{
\item{dbInfo}{Database info}

\item{llmReview}{Output of the llm_review() function}

\item{commit}{(Default = T)}
}
\value{
A data frame with the following columns:
\itemize{
\item evaluation_id,
\item review_assignment_id,
\item review_score_id: the ID for the each detected competency review scores
}
}
\description{
Add an AI response from llm_review() to the database
}
