% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbReviewAssignment}
\alias{dbReviewAssignment}
\title{Insert or update a review assignment}
\usage{
dbReviewAssignment(
  dbInfo,
  id,
  reviewer_id,
  evaluation_id,
  review_prompt_id,
  include_questions,
  redacted,
  duration,
  statusCode,
  tokens_in,
  tokens_out,
  note,
  timestamp,
  commit = T
)
}
\arguments{
\item{dbInfo}{dbInfo object}

\item{id}{(Optional) Review assignment ID. If not set, new entry is created}

\item{reviewer_id}{(Required if id not set)}

\item{evaluation_id}{(Required if id not set)}

\item{review_prompt_id}{(Required if id not set)}

\item{include_questions}{(Optional value)}

\item{redacted}{(Optional value)}

\item{duration}{(Optional value)}

\item{statusCode}{(Optional value)}

\item{tokens_in}{(Optional value)}

\item{tokens_out}{(Optional value)}

\item{note}{(Optional value)}

\item{timestamp}{(Optional value)}

\item{commit}{(Default = T)}
}
\value{
A data frame with inserted / updated database records in review_assignment table
}
\description{
Insert or update a review assignment
}
