% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbReviewer}
\alias{dbReviewer}
\title{Internal function to insert or update into reviewer table}
\usage{
dbReviewer(conn, data, commit = T)
}
\arguments{
\item{conn}{SQLite connection}

\item{data}{Data frame with table columns}

\item{commit}{(Default = TRUE) Commit the transaction}
}
\value{
Inserted / Updated data frame
}
\description{
Internal function to insert or update into reviewer table
}
