% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbReviewerAI}
\alias{dbReviewerAI}
\title{Insert or Update AI reviewer info into the database}
\usage{
dbReviewerAI(conn, id, model, note, commit = T)
}
\arguments{
\item{conn}{CFME database connection}

\item{id}{(Optional) Reviewer id. If provided this means updating existing.
If not, a new reviewer will be created}

\item{model}{AI model name. Required if new reviewer}

\item{note}{(Optional) Text note}

\item{commit}{(Default = T) Commit the changes to the database}
}
\value{
Data frame with inserted / updated reviewer info
}
\description{
Insert or Update AI reviewer info into the database
}
