% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbReviewScore}
\alias{dbReviewScore}
\title{Insert or update into review score table}
\usage{
dbReviewScore(dbInfo, scores, reviewStatus, commit = T)
}
\arguments{
\item{dbInfo}{dbInfo object}

\item{scores}{Data frame with scores}

\item{reviewStatus}{(Optional) Set the review status}

\item{commit}{(Default = T) Commit to DB}
}
\value{
Data frame with changed reviews
}
\description{
Insert or update into review score table
}
