% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getFunArgs}
\alias{getFunArgs}
\title{Get the set (function) arguments of the current environment}
\usage{
getFunArgs(exclude)
}
\value{
A list with the set function arguments and their values
}
\description{
This is useful at the start of a function to capture all passed arguments
}
