% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_chat_completion}
\alias{llm_chat_completion}
\title{Call an Azure chat completion LLM}
\usage{
llm_chat_completion(
  user,
  system,
  log,
  model = "gpt-4o-1120",
  maxTokens = 500,
  version = "2024-10-21",
  endpoint = "https://azure-ai.hms.edu"
)
}
\arguments{
\item{user}{User prompt}

\item{system}{Default = "You are a helpful AI assistant". System prompt}

\item{log}{If set, the token usage is kept track of in this CSV file}

\item{model}{Default = gpt-4o-1120. LLM model to use}

\item{maxTokens}{Default = 500. Max Number of tokens to return}

\item{version}{Default = 2024-10-21.}

\item{endpoint}{Default = https://azure-ai.hms.edu. Azure endpoint}
}
\value{
LLM Response object (list).

If a log file is et a line is added to a CSV file with columns
timestamp, promptTokens and responseTokens (headers not included)
}
\description{
Note that this function expects an environment variable HMS_AZURE_API
that contains the API key You can set this up using
\code{Sys.setenv(HMS_AZURE_API = "API token here")}
}
