% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_responses}
\alias{llm_responses}
\title{Call an Azure response LLM}
\usage{
llm_responses(
  input,
  instructions,
  log,
  model = "gpt-5-mini",
  endpoint = "https://azure-ai.hms.edu"
)
}
\arguments{
\item{input}{User prompt}

\item{instructions}{Default = "You are a helpful AI assistant". System prompt}

\item{log}{If set, the token usage is kept track of in this CSV file}

\item{model}{Default = gpt-4o-1120. LLM model to use}

\item{endpoint}{Default = https://azure-ai.hms.edu. Azure endpoint}
}
\value{
LLM Response object (list).

If a log file is et a line is added to a CSV file with columns
timestamp, promptTokens and responseTokens (headers not included)
}
\description{
Note that this function expects an environment variable HMS_AZURE_API
that contains the API key You can set this up using
\code{Sys.setenv(HMS_AZURE_API = "API token here")}
}
