% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{missingVal}
\alias{missingVal}
\title{Provide missing values if variable does not exist}
\usage{
missingVal(var, useNull = F, n = 1)
}
\arguments{
\item{var}{Variabe to check}

\item{useNull}{(Default = F). Return NA if FALSE else NULL}

\item{n}{(Default = 1) How may times to repeat NA}
}
\value{
A vector of values, NAs or NULL depending on settings
}
\description{
Provide missing values if variable does not exist
}
