% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_highlight.R
\name{mod_highlight_server}
\alias{mod_highlight_server}
\title{Module server for text highlights}
\usage{
mod_highlight_server(id, defaults, reset = reactiveVal())
}
\arguments{
\item{id}{Module ID}

\item{defaults}{A (reactive) vector of default list options}

\item{reset}{An optional reactive variable that will rest to the defaults when triggered}
}
\value{
A reactive data frame with selected pieces of text and an ID in the
order they were listed on the page
}
\description{
Module server for text highlights
}
