% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_setup.R
\name{mod_setup_server}
\alias{mod_setup_server}
\title{Module server setup tab}
\usage{
mod_setup_server(id, conn)
}
\arguments{
\item{id}{Module ID}

\item{conn}{connection to the database}
}
\value{
trigger if there is an update
}
\description{
Module server setup tab
}
