% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{parsePrompt}
\alias{parsePrompt}
\title{Check if a prompt is structured correctly and returned a parsed version}
\usage{
parsePrompt(prompt)
}
\arguments{
\item{prompt}{String of text to check}
}
\value{
list with
\itemize{
\item success: TRUE / FALSE
\item msg: message
\item content: list containing parsed prompt if successful
}
}
\description{
Check if a prompt is structured correctly and returned a parsed version
}
