% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{pin_dev_get}
\alias{pin_dev_get}
\title{Get a pin}
\usage{
pin_dev_get(pinName, path, tempBackup = T)
}
\arguments{
\item{pinName}{name of the pin to access}

\item{path}{Path to save the file to}
}
\value{
list with new file and temp backup if set
}
\description{
Get a pin
}
