% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{pin_dev_set}
\alias{pin_dev_set}
\title{Set a pin}
\usage{
pin_dev_set(pinName, path, nBackups = 3)
}
\arguments{
\item{pinName}{Name of the pin}

\item{path}{Path to the file to pin}

\item{nBackups}{(Default = 3) N most recent pins to keep online}
}
\value{
The name of the new pin
}
\description{
Set a pin
}
