% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualReview.R
\name{reviewDoc}
\alias{reviewDoc}
\title{Generate a document for manual review of evaluations}
\usage{
reviewDoc(
  path,
  dbInfo,
  ids,
  n,
  completeOnly = F,
  redacted = T,
  includeQuestions = T,
  html = F
)
}
\arguments{
\item{path}{Path the file where to write the output (html or txt)}

\item{dbInfo}{Database info}

\item{ids}{IDs to pull, if empty, n must be provided}

\item{n}{Ignored if ids is set, otherwise number of random evaluations to pull}

\item{completeOnly}{(Default = FALSE) Only include complete reviews}

\item{redacted}{(Default = TRUE) Show redacted text}

\item{includeQuestions}{(Default = TRUE) Add the questions to the text}

\item{html}{(Default = FALSE) Output HTML instead of plain text}
}
\value{
A file is written to the specified path and a data frame with
all evaluations is returned
}
\description{
Generate a document for manual review of evaluations
}
