% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbAddPrompt}
\alias{dbAddPrompt}
\title{Add a new prompt to the database}
\usage{
dbAddPrompt(prompt, dbInfo, note, showWarning = T)
}
\arguments{
\item{prompt}{Single string of system prompt text}

\item{dbInfo}{DB connection info}

\item{note}{(Optional) Note about this prompt}

\item{showWarning}{(Default = TRUE) Show warning if prompt already exists}
}
\value{
Prompt ID
}
\description{
Add a new prompt to the database
}
