% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbOperations.R
\name{dbReviewerHuman}
\alias{dbReviewerHuman}
\title{Insert or Update human reviewer info into the database}
\usage{
dbReviewerHuman(conn, id, username, first, last, note, commit = T)
}
\arguments{
\item{conn}{CFME database connection}

\item{id}{(Optional) Reviewer id. If provided this means updating existing.
If not, a new reviewer will be created}

\item{username}{Username. Required if new reviewer}

\item{first}{(Optional) first name}

\item{last}{(Optional) last name}

\item{note}{(Optional) note}

\item{commit}{(Default = T) Commit the changes to the database}
}
\value{
Data frame with inserted / updated reviewer info
}
\description{
Insert or Update human reviewer info into the database
}
