% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{llm_review}
\alias{llm_review}
\title{Get the data from an LLM evaluation}
\usage{
llm_review(dbInfo, review_assignment_id, log, maxTries = 3)
}
\arguments{
\item{dbInfo}{Database info}

\item{review_assignment_id}{IDs in review_assignment table}

\item{log}{(Optional) Save token usage to extra file (will also be in database)}

\item{maxTries}{(Default = 3) How many times to try in case the response is
not in a valid format}
}
\value{
A list with results for each review_assignment_id
\itemize{
\item statusCode = final status code (3 = success)
\item review_assignment_id: ID from review_assignment table in database
\item data: If successful, data frame with response, otherwise NULL
\item tries: Number of times tried for valid response
\item tokens in / out: token used
}
}
\description{
Get the data from an LLM evaluation
}
